requirements.txt
pandas
requests
python-dotenv
pytz
TA-Lib
